viewで使わないメソッドをhelperからcontrollerに移譲
